Moving IForOfState interface in the for_of.directive file and removing unused imports
