--- projects/igniteui-angular/src/lib/directives/for-of/IForOfState.ts
@@ -1,4 +1,0 @@
-export declare interface IForOfState {
-    startIndex?: number;
-    chunkSize?: number;
-}
