+++ projects/igniteui-angular/src/lib/grid/grid.initialization.spec.ts
@@ -5,7 +5,7 @@ import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { BehaviorSubject, Observable } from 'rxjs';
 
 import { DataType } from '../data-operations/data-util';
-import { IForOfState} from '../directives/for-of/IForOfState';
+import { IForOfState} from '../directives/for-of/for-of.directive';
 import { IgxColumnComponent } from './column.component';
 import { IgxGridComponent } from './grid.component';
 import { IgxGridModule } from './index';
